No Hack No CTF 2025/🦢 Let's Cook Some Delicious Goose!/server/requirements.txt
grpcio==1.63.0
grpcio-tools==1.63.0
requests==2.32.3
flask==3.1.1
